+++ src/GSvar/DBTableWidget.h
@@ -73,7 +73,6 @@ signals:
 
 protected:
 	void keyPressEvent(QKeyEvent* event) override;
-	void copyToClipboard(bool selection_only);
 
 protected slots:
 	void copySelectionToClipboard();