+++ doc/GSvar/changelog.md
@@ -1,5 +1,10 @@
 # GSvar change log
 
+### 2021_06-75 (11.08.21)
+- Added ClinVar upload of variants, removed LOVD upload.
+- Added option to show BAF histogram.
+- The repeat expansion table can be copied to the clipboard via the context menu now.
+
 ### 2021_06-57 (03.08.21)
 - added cohort analysis dialog (NGSD > Variants > Cohort analysis)
 - fixed DGV links (no https)